컨트롤러 리팩토링

+++ app/Http/Controllers/Reviews/ReviewController.php
@@ -1,32 +1,37 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Reviews;
 
+use App\Http\Controllers\Controller;
 use App\Http\Requests\Review\CreateReviewRequest;
 use App\Http\Requests\Review\DeleteReviewRequest;
 use App\Http\Requests\Review\ListReviewRequest;
 use App\Http\Requests\Review\UpdateReviewRequest;
 use App\Transformers\ReviewTransformer;
 use Myshop\Application\Service\ReviewService;
-use Myshop\Domain\Model\Product;
-use Myshop\Domain\Model\Review;
+use Myshop\Domain\Repository\ProductRepository;
 use Myshop\Domain\Repository\ReviewRepository;
 
 class ReviewController extends Controller
 {
-    private $reviewService;
+    private $productRepository;
     private $reviewRepository;
+    private $reviewService;
 
     public function __construct(
-        ReviewService $reviewService,
-        ReviewRepository $reviewRepository
+        ProductRepository $productRepository,
+        ReviewRepository $reviewRepository,
+        ReviewService $reviewService
     ) {
+        $this->productRepository = $productRepository;
         $this->reviewService = $reviewService;
         $this->reviewRepository = $reviewRepository;
     }
 
-    public function index(ListReviewRequest $request, Product $product)
+    public function index(ListReviewRequest $request, int $productId)
     {
+        $product = $this->productRepository->findById($productId);
+
         $paginatedReviewCollection = $this->reviewRepository->findBySearchParam(
             $request->getReviewSearchParam(), $product
         );
@@ -37,8 +42,10 @@ class ReviewController extends Controller
         );
     }
 
-    public function store(CreateReviewRequest $request, Product $product)
+    public function store(CreateReviewRequest $request, int $productId)
     {
+        $product = $this->productRepository->findById($productId);
+
         $review = $this->reviewService->createReview(
             $product, $request->user(), $request->getReviewDto()
         );
@@ -48,9 +55,11 @@ class ReviewController extends Controller
 
     public function update(
         UpdateReviewRequest $request,
-        Product $product,
+        int $productId,
         int $reviewId
     ) {
+        $product = $this->productRepository->findById($productId);
+
         // [선점잠금] 레코드를 조회하고 잠급니다.
         // $review = $this->reviewRepository->findByIdWithLock($reviewId, $product);
 
@@ -62,6 +71,8 @@ class ReviewController extends Controller
         // 조회시점 대비 DB의 버전이 같은지를 확인하여 변경 가능 여부를 판단합니다.
         $review = $this->reviewRepository->findById($reviewId, $product);
 
+        $this->authorize('update', $review);
+
         $review = $this->reviewService->modifyReview(
             $review, $request->getReviewDto()
         );
@@ -71,9 +82,14 @@ class ReviewController extends Controller
 
     public function destroy(
         DeleteReviewRequest $request,
-        Product $product,
-        Review $review
+        int $productId,
+        int $reviewId
     ) {
+        $product = $this->productRepository->findById($productId);
+        $review = $this->reviewRepository->findById($reviewId);
+
+        $this->authorize('delete', $review);
+
         $this->reviewService->deleteReview($review, $product);
 
         return json()->noContent();
